// src/util/types/Door.ts
/**
 * src/util/types/Door.ts
 *
 * Author: Ben Siebert <hello@ben-siebert.de>
 * Copyright: Copyright (c) 2018-2023 Ben Siebert. All rights reserved.
 * License: Project License
 * Created At: 13.11.2023
 *
 */

export type Door = {
  calendar: string;
  day: number;
  title: string;
  description: string;
  image: string;
  _id: string;
};